fix warnings

+++ test/core/executor.cpp
@@ -39,7 +39,7 @@ TEST_CASE("Executor Equality")
 TEST_CASE("Can use CPUExecutor wo mempool")
 {
     auto exec = NeoN::CPUExecutor();
-    int numElems {5};
+    size_t numElems {5};
     int* ptr = exec.template alloc<int>(numElems);
     exec.free(ptr);
 }
@@ -64,7 +64,7 @@ TEST_CASE("Can use DefaultExecutor wo mempool")
 TEST_CASE("Can use SerialExecutor with UmpireAllocator")
 {
     auto exec = NeoN::SerialExecutor(std::make_unique<NeoN::UmpireAllocator>());
-    int numElems {5};
+    size_t numElems {5};
     int* ptr = exec.template alloc<int>(numElems);
     exec.free(ptr);
 }
@@ -73,7 +73,7 @@ TEST_CASE("Can use SerialExecutor with UmpireAllocator")
 TEST_CASE("Can use DefaultExecutor with UmpireAllocator")
 {
     auto exec = NeoN::createDefaultExecutor(std::make_unique<NeoN::UmpireAllocator>());
-    int numElems {5};
+    size_t numElems {5};
     std::visit(
         [=](auto concreteExec)
         {
@@ -90,7 +90,7 @@ TEST_CASE("Can use SerialExecutor with UmpirePoolAllocator")
     auto exec = NeoN::SerialExecutor(std::make_unique<NeoN::UmpirePoolAllocator>());
     NeoN::UmpireMempoolHandler::setupUmpirePool(NeoN::MemorySpace::CPU, 1024);
 
-    int numElems {5};
+    size_t numElems {5};
     int* ptr = exec.template alloc<int>(numElems);
     exec.free(ptr);
 
@@ -103,7 +103,7 @@ TEST_CASE("Can use DefaultExecutor with UmpirePoolAllocator")
     auto exec = NeoN::createDefaultExecutor(std::make_unique<NeoN::UmpirePoolAllocator>());
     NeoN::UmpireMempoolHandler::setupUmpirePool(NeoN::memorySpace(exec), 1024);
 
-    int numElems {5};
+    size_t numElems {5};
     std::visit(
         [=](auto concreteExec)
         {
